FEAT: Add feature use_std; support no_std mode
Split out affected adaptors to their own files
