+++ .travis.yml
@@ -13,6 +13,7 @@ branches:
     - master
 script:
   - |
+      cargo build --verbose --no-default-features &&
       cargo build --verbose --features "$FEATURES" &&
       cargo test --verbose --features "$FEATURES" &&
       ([ "$BENCH" != 1 ] || cargo bench --verbose --features "$FEATURES")
