+++ src/multipeek_impl.rs
